+++ app/layout.tsx
@@ -9,7 +9,7 @@ import classnames from 'classnames';
 import Footer from '@/components/Common/footer/Footer';
 import Header from '@/components/Common/header/Header';
 import Provider from '@/components/Common/Provider';
-import './globals.scss';
+import '@/styles/globals.scss';
 import '@mantine/spotlight/styles.css';
 import styles from './layout.module.scss';
